Memory optimizations

+++ server/serialization/MSECParser.cs
@@ -103,14 +103,14 @@ namespace Maps.Serialization
                         Route route = new Route();
 
                         if (Regex.IsMatch(tokens[cur], @"^[-+]?[01]$"))
-                            route.StartOffsetX = Int32.Parse(tokens[cur++], NumberStyles.Integer, CultureInfo.InvariantCulture);
+                            route.StartOffsetX = sbyte.Parse(tokens[cur++], NumberStyles.Integer, CultureInfo.InvariantCulture);
                         if (Regex.IsMatch(tokens[cur], @"^[-+]?[01]$"))
-                            route.StartOffsetY = Int32.Parse(tokens[cur++], NumberStyles.Integer, CultureInfo.InvariantCulture);
+                            route.StartOffsetY = sbyte.Parse(tokens[cur++], NumberStyles.Integer, CultureInfo.InvariantCulture);
                         route.Start = Astrometrics.HexToInt(tokens[cur++]);
                         if (Regex.IsMatch(tokens[cur], @"^[-+]?[01]$"))
-                            route.EndOffsetX = Int32.Parse(tokens[cur++], NumberStyles.Integer, CultureInfo.InvariantCulture);
+                            route.EndOffsetX = sbyte.Parse(tokens[cur++], NumberStyles.Integer, CultureInfo.InvariantCulture);
                         if (Regex.IsMatch(tokens[cur], @"^[-+]?[01]$"))
-                            route.EndOffsetY = Int32.Parse(tokens[cur++], NumberStyles.Integer, CultureInfo.InvariantCulture);
+                            route.EndOffsetY = sbyte.Parse(tokens[cur++], NumberStyles.Integer, CultureInfo.InvariantCulture);
                         route.End = Astrometrics.HexToInt(tokens[cur++]);
                         if (cur < tokens.Length)
                             route.ColorHtml = tokens[cur++];
